base/requirements.txt
snowflake-snowpark-python
snowflake-cli
python-gnupg
python-dotenv